Added the closest station distance computation. Based on it, LandmarksActivity now gets the places around that distance.

+++ app/src/main/java/co/swatisi/team/metrowalkdc/model/Station.kt
@@ -1,3 +1,3 @@
 package co.swatisi.team.metrowalkdc.model
 
-data class Station(val name: String, val lat: Float, val lon: Float, val lineCode1: String)
+data class Station(val name: String, val lat: Double, val lon: Double, val lineCode1: String)
